% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qframe.R
\name{as.qframe}
\alias{as.qframe}
\title{Creates a \code{\link{qframe}} from a \code{\link[base]{data.frame}}.}
\usage{
as.qframe(df)
}
\arguments{
\item{df}{the \code{\link[base]{data.frame}} to convert}
}
\description{
Creates a \code{\link{qframe}} from a \code{\link[base]{data.frame}}.
}
