% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qframe.R
\name{[.qframe}
\alias{[.qframe}
\title{Extract values from a \code{\link{qframe}}.}
\usage{
\method{[}{qframe}(x, i, j, drop)
}
\arguments{
\item{x}{the qframe to extract data from}

\item{i}{elements to extract or replace, if a character vector, uses \code{\link[stringr]{str_detect}} to collect the elements}

\item{j}{elements to extract or replace, if a character vector, uses \code{\link[stringr]{str_detect}} to collect the elements}

\item{drop}{see \code{\link[base]{data.frame}}}
}
\description{
Extract values from a \code{\link{qframe}}.
}
