% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qframe.R
\name{[<-.qframe}
\alias{[<-.qframe}
\title{Replace values from a \code{\link{qframe}}.}
\usage{
\method{[}{qframe}(x, i, j) <- value
}
\arguments{
\item{x}{the qframe to extract data from}

\item{i}{elements to extract or replace, if a character vector, uses \code{\link[stringr]{str_detect}} to collect the elements, see \code{\link[base]{data.frame}}}

\item{j}{elements to extract or replace, if a character vector, uses \code{\link[stringr]{str_detect}} to collect the elements, see \code{\link[base]{data.frame}}}

\item{value}{a suitable replacement value, see \code{\link[base]{data.frame}}}
}
\description{
Replace values from a \code{\link{qframe}}.
}
